feat:marca - Added status code HTTP

+++ app/Http/Controllers/MarcaController.php
@@ -23,7 +23,7 @@ class MarcaController extends Controller
     {
        // $marca = Marca::all();
         $marcas = $this->marca->all();
-        return $marcas;
+        return response()->json($marcas,200);
     }
 
     /**
@@ -36,7 +36,7 @@ class MarcaController extends Controller
     {
         //$marca = Marca::create($request->all());
         $marca = $this->marca->create($request->all());
-        return $marca;
+        return response()->json($marca,201);
     }
 
     /**
@@ -48,7 +48,10 @@ class MarcaController extends Controller
     public function show($id)
     {
         $marca = $this->marca->find($id);
-        return $marca;
+        if($marca === null){
+            return response()->json(['erro'=>'Recurso pesquisado não existe'],404);
+        }
+        return response()->json($marca,200);
     }
 
     /**
@@ -65,8 +68,11 @@ class MarcaController extends Controller
         // print_r($marca->getAttributes());
         // $marca->update($request->all());
         $marca = $this->marca->find($id);
+        if($marca === null){
+            return response()->json(['erro'=>'Atualização não realizada. Recurso pesquisado não existe', 404]);
+        }
         $marca->update($request->all());
-        return $marca;
+        return response()->json($marca,200);
     }
 
     /**
@@ -78,7 +84,10 @@ class MarcaController extends Controller
     public function destroy($id)
     {
         $marca = $this->marca->find($id);
+        if($marca === null){
+            return response()->json(['erro'=>'Exclusão  não realizada. Recurso pesquisado não existe'], 404);
+        }
         $marca->delete();
-        return ['msg', 'A marca foi deletada com sucesso!'];
+        return response()->json(['msg', 'A marca foi deletada com sucesso!'],200);
     }
 }
